feat: renames files & adds design tokens
