+++ .gitattribute
@@ -1,0 +1,46 @@
+* text=auto
+
+*.js text eol=lf
+*.cjs text eol=lf
+*.ejs text eol=lf
+*.mjs text eol=lf
+*.ts text eol=lf
+*.json text eol=lf
+*.jsonc text eol=lf
+*.json5 text eol=lf
+@@ -14,11 +10,7 @@
+*.yaml text eol=lf
+*.yml text eol=lf
+*.xml text eol=lf
+*.svg text eol=lf
+*.txt text eol=lf
+*.ini text eol=lf
+*.md text eol=lf
+*.html text eol=lf
+*.css text eol=lf
+*.properties text eol=lf
+*.webmanifest text eol=lf
+
+*.png binary
+*.webp binary
+*.avif binary
+*.jpg binary
+*.gif binary
+*.ico binary
+*.jpeg binary
+@@ -32,12 +24,8 @@
+*.otf binary
+*.zip binary
+*.crx binary
+*.woff binary
+*.woff2 binary
+*.ttf binary
+*.eot binary
+*.pdf binary
+*.mp4 binary
+*.mp3 binary
+*.wav binary
+
+.gitattributes export-ignore
+.gitignore export-ignore
+.github/ export-ignore
