--- .gitattributes
@@ -1,44 +1,0 @@
-# Set default behavior for all files
-* text=auto
-
-# Explicitly declare text files you want to always be normalized and converted
-# to native line endings on checkout
-*.js text eol=lf
-*.ts text eol=lf
-*.json text eol=lf
-*.html text eol=lf
-*.css text eol=lf
-*.scss text eol=lf
-*.md text eol=lf
-*.yml text eol=lf
-*.yaml text eol=lf
-*.xml text eol=lf
-*.txt text eol=lf
-
-# Declare files that will always have CRLF line endings on checkout
-*.bat text eol=crlf
-
-# Denote all files that should not be normalized and should be left as-is
-*.png binary
-*.jpg binary
-*.jpeg binary
-*.gif binary
-*.ico binary
-*.svg binary
-*.woff binary
-*.woff2 binary
-*.eot binary
-*.ttf binary
-*.otf binary
-*.zip binary
-*.crx binary
-
-# Extension-specific files
-manifest.json text eol=lf
-*.webmanifest text eol=lf
-
-# Export-ignore files for git archive
-.gitattributes export-ignore
-.gitignore export-ignore
-.github/ export-ignore
-tests/ export-ignore
