+++ .npmrc
@@ -1,0 +1,14 @@
+registry = https://registry.npmjs.org/
+strict-ssl = true
+save-prefix =
+
+strict-peer-deps = true
+prefer-dedupe = true
+save-exact = true
+
+audit = true
+audit-level = critical
+package-lock = true
+
+progress = true
+update-notifier = true
