+++ shared/constants.js
@@ -250,7 +250,7 @@ export const ELEMENT_IDS = {
   JIRA_TICKETS: 'jira-tickets',
   GITLAB_HISTORY: 'gitlab-history',
   OPTIONS_LINK: 'optionsLink',
-  
+
   VIEW_DEMO_REPORT_DETAILS: 'viewDemoReportDetails',
   EXAMPLE_RESULTS: 'exampleResults',
   JIRA_TICKETS_EXAMPLE: 'jira-tickets-example',
