--- views/extensionSidePanel.html
@@ -1,205 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <meta
-      name="viewport"
-      content="width=device-width, initial-scale=1, viewport-fit=cover"
-    />
-    <title data-i18n="toolInterface">Tool Interface</title>
-    <link href="../shared/designTokens.css" rel="stylesheet" />
-    <script
-      defer
-      src="../helpers/internationalizationHelper.js"
-      type="module"
-    ></script>
-    <script
-      defer
-      src="../services/sidePanelApiManager.js"
-      type="module"
-    ></script>
-  </head>
-  <body style="overflow-x: auto; position: relative">
-    <div class="container" style="min-width: 42.375rem">
-      <header class="header-section">
-        <p role="note" class="info info-notice">
-          <small data-i18n="compareNotice"
-            >Compare the user stories and git commit history associated with a
-            release to see the differences between <em>Jira tickets</em>
-            <abbr title="Verses" data-i18n-title="versesAbbr">&</abbr>
-            <em>GitLab changes</em>.</small
-          >
-        </p>
-        <h1
-          style="
-            display: flex;
-            justify-content: space-between;
-            align-items: center;
-            margin-top: 1.5rem;
-          "
-        >
-          <span data-i18n="releaseOverview">Release Overview</span>
-          <a
-            href="./userConfigurations.html"
-            id="optionsLink"
-            title="Settings"
-            data-i18n-title="settings"
-            target="_blank"
-            rel="noopener noreferrer"
-            style="font-size: 1.5rem; text-decoration: none"
-            >⚙️</a
-          >
-        </h1>
-      </header>
-      <main role="main">
-        <div class="grid grid-2-col">
-          <!-- Jira Fields -->
-          <div class="form-section">
-            <div class="form-group">
-              <label for="jiraProjectKey" data-i18n="jiraProjectKey"
-                >Jira Project Key</label
-              >
-              <input
-                type="text"
-                id="jiraProjectKey"
-                placeholder="e.g. DDSTM"
-                data-i18n-placeholder="jiraProjectKeyPlaceholder"
-                value="DDSTM"
-              />
-            </div>
-          </div>
-          <!-- GitLab Fields -->
-          <div class="form-section">
-            <div class="form-group">
-              <label for="gitlabProjectId" data-i18n="gitlabProjectId"
-                >GitLab Project ID</label
-              >
-              <input
-                type="text"
-                id="gitlabProjectId"
-                placeholder="e.g. #####"
-                data-i18n-placeholder="gitlabProjectIdPlaceholder"
-              />
-              <p class="input-hint" data-i18n="gitlabProjectIdHint">
-                Find in: <samp>Settings > General > Project ID</samp>
-              </p>
-            </div>
-          </div>
-        </div>
-        <br />
-        <div class="grid grid-2-col flex align-end">
-          <div class="form-section">
-            <div class="form-group">
-              <label for="jiraFixVersion" data-i18n="jiraReleaseVersionId"
-                >Jira Release Version ID</label
-              >
-              <input
-                type="text"
-                id="jiraFixVersion"
-                placeholder="e.g. #####"
-                data-i18n-placeholder="jiraReleaseVersionIdPlaceholder"
-                list="versionsDatalist"
-              />
-              <p class="input-hint" data-i18n="jiraReleaseVersionIdHint">
-                Find in the release URL:
-                <samp>.../browse/DDSTM/1.1.0/#####</samp>
-              </p>
-              <button
-                id="getVersionsBtn"
-                class="btn btn-secondary btn-sm"
-                data-i18n="getReleases"
-              >
-                Get Releases
-              </button>
-              <datalist id="versionsDatalist"></datalist>
-            </div>
-          </div>
-          <div class="form-section">
-            <div class="form-group">
-              <label for="gitlabCurrentTag" data-i18n="gitlabCurrentReleaseTag"
-                >GitLab Current Release Tag</label
-              >
-              <input
-                type="text"
-                id="gitlabCurrentTag"
-                placeholder="e.g. v1.0.1"
-                data-i18n-placeholder="gitlabCurrentReleaseTagPlaceholder"
-              />
-            </div>
-            <div class="form-group">
-              <label
-                for="gitlabPreviousTag"
-                data-i18n="gitlabPreviousReleaseTag"
-                >GitLab Previous Release Tag</label
-              >
-              <input
-                type="text"
-                id="gitlabPreviousTag"
-                placeholder="e.g. v1.0.0"
-                data-i18n-placeholder="gitlabPreviousReleaseTagPlaceholder"
-              />
-            </div>
-          </div>
-        </div>
-        <br />
-        <div class="grid grid-2-col" style="margin-bottom: 1rem;">
-          <button id="resetBtn" class="btn btn-secondary" data-i18n="reset">
-            Reset
-          </button>
-          <button
-            id="generateSummaryBtn"
-            class="btn btn-primary"
-            data-i18n="generateSummary"
-          >
-            Generate Release Report
-          </button>
-        </div>
-        <details id="viewDemoReportDetails" style="margin-bottom: 1rem;">
-          <summary data-i18n="viewDemoReport" style="cursor: pointer; padding: 0.5rem; border: 1px solid var(--border-color); border-radius: var(--border-radius); background-color: var(--surface-color);">View Demo v2.26.0 Release Data</summary>
-          <div id="exampleResults" class="results-section" style="margin-top: 0.5rem;padding-top: 1rem;">
-            <!--  Demo Report Sections will be populated here by JS  -->
-            <div class="side-by-side">
-              <div class="column report-section">
-                <h3 data-i18n="jiraTickets">Jira Tickets (Demo)</h3>
-                <div id="jira-tickets-example">
-                  <!--  Populated by JS  -->
-                </div>
-              </div>
-              <div class="column report-section">
-                <h3 data-i18n="gitlabCommits">GitLab Commits (Demo)</h3>
-                <div id="gitlab-commits-example">
-                  <!--  Populated by JS  -->
-                </div>
-              </div>
-            </div>
-            <div class="report-section" style="margin-top: 1rem;">
-              <h3 data-i18n="gitlabTags">GitLab Tags (Demo)</h3>
-              <div id="gitlab-tags-example">
-                <!--  Populated by JS  -->
-              </div>
-            </div>
-          </div>
-        </details>
-        <div id="loading" class="loading-spinner hidden"></div>
-        <div id="statusMessage" class="status-message"></div>
-        <div id="summaryResults" class="results-section hidden">
-          <div class="side-by-side">
-            <div class="column">
-              <h2 data-i18n="jiraTickets">Jira Tickets</h2>
-              <div id="jira-tickets"></div>
-            </div>
-            <div class="column">
-              <h2 data-i18n="gitlabHistory">GitLab History</h2>
-              <div id="gitlab-history"></div>
-            </div>
-          </div>
-        </div>
-      </main>
-      <footer class="footer-section">
-        <small data-i18n="poweredBy"
-          >Powered by the <a href="https://developer.chrome.com/docs/extensions/reference/api" title="View Documentation" aria-label="Visit Google's Chrome Extension API specification page">Chrome Extension API</a></small
-        >
-      </footer>
-    </div>
-  </body>
-</html>
