+++ views/options.html
@@ -1,0 +1,117 @@
+<!doctype html>
+<html lang="en">
+
+<head>
+  <meta charset="utf-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1, viewport-fit=cover">
+  <title data-i18n="toolSettings">Release Diff Tool Settings</title>
+  <link href="../shared/styles.css" rel="stylesheet">
+  <script defer src="../utils/i18n.util.js" type="module"></script>
+  <script defer src="../services/api/options.api.js" type="module"></script>
+</head>
+
+<body style="overflow-x: auto; position: relative;">
+  <div class="container" style="min-width: 52.8125rem;">
+    <div role="alert" class="alert alert-warning" style="margin-top: .85rem;">
+      <small>
+        <strong data-i18n="securityNoticeTitle">Important Security Notice:</strong> <span
+          data-i18n="securityNoticeText">For
+          enterprise production environments, storing <abbr title="Application Programming Interface"
+            data-i18n-title="apiAbbr">API</abbr> Personal
+          Access Tokens (<abbr title="Personal Access Token" data-i18n-title="patAbbr">PAT</abbr>s) directly
+          in the browser's
+          client storage
+          is
+          generally
+          <ins>NOT</ins>
+          recommended due to
+          security risks. A backend proxy service is the more secure approach for handling credentials.
+          <em>This
+            extension uses <code>localStorage</code> currently.</em></span>
+      </small>
+    </div>
+    <header class="header-section">
+      <h1 data-i18n="extensionSettings">Extension Settings</h1>
+      <div class="form-group form-group-no-margin">
+        <label for="themeToggle" data-i18n="themeToggle">Theme:</label>
+        <input type="checkbox" id="themeToggle" name="themeToggle">
+        <span data-i18n="themeToggleDark">Dark Mode</span>
+      </div>
+    </header>
+    <main role="main">
+      <div class="grid grid-2-col">
+        <div class="form-section">
+          <div class="form-group">
+            <label for="jiraBaseUrl" data-i18n="jiraUrl">Jira <abbr title="Uniform Resource Locator"
+                data-i18n-title="urlAbbr">URL</abbr></label>
+            <input type="url" id="jiraBaseUrl" placeholder="e.g. https://jira.atlassian.com"
+              data-i18n-placeholder="jiraUrlPlaceholder" value="https://jira.dell.com/">
+            <p class="input-hint" data-i18n="jiraUrlHint">e.g. <samp>https://jira.dell.net</samp></p>
+          </div>
+        </div>
+        <div class="form-section">
+          <div class="form-group">
+            <label for="gitlabBaseUrl" data-i18n="gitlabUrl">GitLab <abbr title="Uniform Resource Locator"
+                data-i18n-title="urlAbbr">URL</abbr></label>
+            <input type="url" id="gitlabBaseUrl" placeholder="e.g. https://gitlab.enterprise.com"
+              data-i18n-placeholder="gitlabUrlPlaceholder" value="https://gitlab.dell.com/">
+            <p class="input-hint" data-i18n="gitlabUrlHint">e.g. <samp>https://gitlab.dell.net</samp></p>
+          </div>
+        </div>
+      </div>
+      <br />
+      <div class="grid grid-2-col">
+        <div class="form-section">
+          <div class="form-group">
+            <label for="jiraPat" data-i18n="jiraPat">Jira <abbr title="Personal Access Token"
+                data-i18n-title="patAbbr">PAT</abbr></label>
+            <input type="password" id="jiraPat" placeholder="provide secret token&hellip;"
+              data-i18n-placeholder="jiraPatPlaceholder">
+            <p class="input-hint" data-i18n="jiraPatHint">Create a <abbr title="Personal Access Token"
+                data-i18n-title="patAbbr">PAT</abbr> in your Atlassian account
+              settings. The
+              token needs read
+              access to projects and issues. <a
+                href="https://support.atlassian.com/atlassian-account/docs/manage-api-tokens-for-your-atlassian-account/"
+                target="_blank">Learn how to generate a Jira <abbr title="Application Programming Interface"
+                  data-i18n-title="apiAbbr">API</abbr>
+                token</a>
+            </p>
+          </div>
+        </div>
+        <div class="form-section">
+          <div class="form-group">
+            <label for="gitlabPat" data-i18n="gitlabPat">GitLab <abbr title="Personal Access Token"
+                data-i18n-title="patAbbr">PAT</abbr></label>
+            <input type="password" id="gitlabPat" placeholder="insert private key&hellip;"
+              data-i18n-placeholder="gitlabPatPlaceholder">
+            <p class="input-hint" data-i18n="gitlabPatHint">Create a <abbr title="Personal Access Token"
+                data-i18n-title="patAbbr">PAT</abbr> in your GitLab profile
+              preferences.
+              The token
+              needs at minimum
+              'read_api' and 'read_repository' scopes. <a
+                href="https://docs.gitlab.com/ee/user/profile/personal_access_tokens.html" target="_blank">Learn how
+                to
+                create a GitLab <abbr title="Personal Access Token" data-i18n-title="patAbbr">PAT</abbr></a>
+            </p>
+          </div>
+        </div>
+      </div>
+    </main>
+    <footer style="margin-top: 3rem">
+      <button id="saveButton" class="btn btn-primary" data-i18n="saveSettings">Save Settings</button>
+      <p id="status" aria-live="polite" aria-atomic="true" aria-relevant="all" class="status-message"></p>
+    </footer>
+    <aside class="footer-section">
+      <div class="btn-group">
+        <button id="testBtn" class="btn btn-secondary" data-i18n="testConnection">Test Connection</button>
+        <button id="testJiraBtn" class="btn btn-secondary" data-i18n="testJiraApi">Test Jira <abbr
+            title="Application Programming Interface" data-i18n-title="apiAbbr">API</abbr></button>
+        <button id="testGitLabBtn" class="btn btn-secondary" data-i18n="testGitlabApi">Test GitLab <abbr
+            title="Application Programming Interface" data-i18n-title="apiAbbr">API</abbr></button>
+      </div>
+    </aside>
+</body>
+
+</html>
